melo_strategic_ai/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary

pydantic-settings
python-dotenv

python-multipart

passlib[bcrypt]
python-jose[cryptography]

PyMuPDF
spacy

loguru
slowapi

stripe

pt-core-news-sm